yesand-sandbox legacy-yesand-sandbox/requirements.txt
aiofiles
fastapi
openai
python-dotenv
uvicorn
